+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+ - Added email opt-in on the registration page [#1210](https://github.com/PublicMapping/districtbuilder/pull/1210)
 
 ### Changed