+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed create-random-projects for large number of states [#1232](https://github.com/PublicMapping/districtbuilder/pull/1232)
+- Fix excess memory usage on projects query [#1228](https://github.com/PublicMapping/districtbuilder/pull/1228)
 
 
 ## [1.17.3]